requirements.txt
click==8.0.3
python_gitlab==4.11.1
bottle==0.13.1
